cmd/snap-fde-keymgr: move the files around in preparation for merging
